Add directed and undirected graphs
Added implementation for undirected graph. Still work in progress.
neighbours function is not fully implemented yet, but there is at least
some result.

Documentation is missing for most of the parts.

Directed graph is not implemented at all. Stubs are done.

This will implemented the missing algorithms mentioned in #3 of TheAlgorithms/Rust